+++ src/basic_recipes/hvspan.jl
@@ -1,10 +1,12 @@
 """
     hspan(ys_low, ys_high; xmin = 0.0, xmax = 1.0, attrs...)
+    hspan(ys_lowhigh; xmin = 0.0, xmax = 1.0, attrs...)
 
 Create horizontal bands spanning across a `Scene` with 2D projection.
 The bands will be placed from `ys_low` to `ys_high` in data coordinates and `xmin` to `xmax`
 in scene coordinates (0 to 1). All four of these can have single or multiple values because
 they are broadcast to calculate the final spans.
+Both bounds can be passed together as an interval `ys_lowhigh`.
 """
 @recipe HSpan begin
     "The start of the bands in relative axis units (0 to 1) along the x dimension."
@@ -17,11 +19,13 @@ end
 
 """
     vspan(xs_low, xs_high; ymin = 0.0, ymax = 1.0, attrs...)
+    vspan(xs_lowhigh; ymin = 0.0, ymax = 1.0, attrs...)
 
 Create vertical bands spanning across a `Scene` with 2D projection.
 The bands will be placed from `xs_low` to `xs_high` in data coordinates and `ymin` to `ymax`
 in scene coordinates (0 to 1). All four of these can have single or multiple values because
 they are broadcast to calculate the final spans.
+Both bounds can be passed together as an interval `xs_lowhigh`.
 """
 @recipe VSpan begin
     "The start of the bands in relative axis units (0 to 1) along the y dimension."
@@ -96,3 +100,5 @@ function data_limits(p::VSpan)
 end
 
 boundingbox(p::Union{HSpan, VSpan}, space::Symbol = :data) = transform_bbox(p, data_limits(p))
+
+convert_arguments(P::Type{<:Union{HSpan, VSpan}}, x::Interval) = convert_arguments(P, endpoints(x)...)